backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

fastapi==0.109.2
uvicorn==0.27.1
python-multipart
opencv-python-headless
numpy
pillow
simple-lama-inpainting
torch==2.5.1+cu121
torchvision==0.20.1+cu121
